% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{generate_pca}
\alias{generate_pca}
\title{PCA plot generation from DESeq2 object using ggplot2}
\usage{
generate_pca(
  deseq2_object,
  transformation = c("vst", "rlog"),
  color_by,
  shape_by = NULL
)
}
\arguments{
\item{deseq2_object}{A DESeq2Results object. Generated for the DESeq2 pipeline.}

\item{transformation}{Character. "vst" or "rlog" depending on user intent & sample size. Blind is set to FALSE.}

\item{color_by}{Character. Variable/annotation by which PCA plot points will be coloured by.}

\item{shape_by}{Character. Optional, variable/annotation by which PCA points will be shaped. Default is NULL.}
}
\value{
PCA plot coloured & shaped by annotations provided.
}
\description{
This function generates a PCA plot from the DESeq2 object by first
applying either a regularised log (rlog, where N < 20) or
variance stabilising tranformation (vst, where N > 20) & then
generating the plot using ggplot2.
}
