% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plots.R
\name{generate_volcano}
\alias{generate_volcano}
\title{Generation of volcano plot from DESeq2 DEGs}
\usage{
generate_volcano(
  res_df,
  padj_cutoff = 0.05,
  lfc_cutoff = 1,
  map_ensembl = TRUE
)
}
\arguments{
\item{res_df}{Data frame. Results from DESeq2 pipeline.}

\item{padj_cutoff}{Numeric. Adjusted p-value threshold. Default is 0.05.}

\item{lfc_cutoff}{Numeric. Log-fold change threshold. Default is 1.}

\item{map_ensembl}{Logical. If TRUE, ENSEMBL IDs will be mapped to gene symbols using AnnotationDbi.}
}
\value{
Volcano plot highlighting the significantly DEGs.
}
\description{
Generates a volcano plot using ggplot using the DEGs output from DESeq2.
Allows for the adjusted p-value cutoff & LFC cutoff to be adjusted, &
whether or not to include ENSEMBL IDs or not.
}
