% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_metadata.R
\name{match_annotations}
\alias{match_annotations}
\title{Read in annotations & match sample names in counts dataframe with metadata}
\usage{
match_annotations(count_data, metadata_path, match_col, clean_col)
}
\arguments{
\item{count_data}{Data frame. The dataframe read in from featureCounts- or HTSeq-generated reads.}

\item{metadata_path}{Character. The path to the metadata .csv file.}

\item{match_col}{Character. The matching column name by which the counts data & metadata will be matched (essentially a barcode to match samples). MUST BE PRESENT IN METADATA FILE.}

\item{clean_col}{Character. The column by which samples in counts data will be renames by (the clean labelling of samples). MUST BE PRESENT IN METADATA FILE.}
}
\value{
A a named list containing both the cleaned up & matched counts data & metadata.
}
\description{
This function works to read in the a metadata .csv file as a dataframe &
-then matches a certain sample name with its associated label & replaces
-a 'messy' label with the clean label (e.g. sample1_trimmed_sorted.txt to sample1_treated).
Finally, the data are sorted to ensure all data is in the same order in both count data frame & metadata.
NOTE: This function uses stringr::str_detect so match_col need to be unique for each sample to act as a unique
identifier for the function, otherwise it will fail & match incorrectly.
}
