% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deseq2.R
\name{print_results}
\alias{print_results}
\title{Exports DESeq2 DEG results in the working directory.}
\usage{
print_results(deseq2_res, file_name)
}
\arguments{
\item{deseq2_res}{A DESeqResults object.}

\item{file_name}{Character. The name of the desired file output (e.g. "results.csv")}
}
\description{
This function orders the DEGs within the dds object & orders it based on the p-value &-
exports the file as provided by the user within the current working directory.
}
