% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data.R
\name{read_featurecounts_data}
\alias{read_featurecounts_data}
\title{Read in & format featureCounts generated counts}
\usage{
read_featurecounts_data(dir_path)
}
\arguments{
\item{dir_path}{Character. Path to .txt file containing featureCounts-generated reads.}
}
\value{
A dataframe containing all raw counts, where rows correspond to gene IDs, & columns correspond to samples.
}
\description{
This function imports the featureCounts output .txt file & formats it by-
-removing any N/A counts, irrelevant metadata columns, & formatting so that-
-rows correspond to gene IDs & columns correspond to samples.
}
