% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_data.R
\name{read_htseq_data}
\alias{read_htseq_data}
\title{Read in & combine all output files from HTSeq}
\usage{
read_htseq_data(dir_path, file_pattern = "\\\\.txt$")
}
\arguments{
\item{dir_path}{Character. Path to folder containing all .txt files from HTSeq output.}

\item{file_pattern}{Character. The regex pattern to match files (default is "\\.txt$")}
}
\value{
A dataframe of raw counts for each gene ID (as rows) & sample (as columns).
}
\description{
This function works to combine all HTSeq output files from a folder of a specified directory.
It combines all files into a single dataframe, setting the row names as the Geneids & removing summary rows at the end.
Allows for data to be ready for downstream analysis
}
