% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_deseq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run the DESeq2 pipeline on subsetted & ordered data & metadata.}
\usage{
run_deseq2(
  count_data,
  metadata,
  design_formula,
  reference_factor,
  reference_level,
  contrast_vector
)
}
\arguments{
\item{count_data}{Data frame. Containing integer values of raw gene counts generated by HTSeq or featureCounts.}

\item{metadata}{Data frame. The matched & ordered metadata .csv file.}

\item{design_formula}{A formula object defining the experimental design (e.g. ~ Age + Treatment)}

\item{reference_factor}{Character. The name of the variable by which data will be re-leveled & compared based on (e.g. Treatment). NEEDS TO BE PRESENT IN METADATA FILE.}

\item{reference_level}{Character. The name of the control condition which serves as the denominator for calculating the DEGs (e.g. PBS vehicle). NEEDS TO BE PRESENT IN METADATA FILE.}

\item{contrast_vector}{Character vector. Definition of the reference factor, the treatment condition & reference level for contrasting & extracting results (e.g. c("Treatment", "Drug 1", "PBS vehicle"))}
}
\value{
A list containing the DESeqDataSet (dds) object & the results table.
}
\description{
This function takes the ordered count data, metadata, along with information regarding-
-the reference factor (e.g. treatment), reference level (e.g. PBS vehicle), & contrast vector-
-(e.g. c("Treatment", "Drug 1", "PBS vehicle)) to run the DESeq2 pipeline.
The results names & summary is printed, & a list containing the dds object & results table is returned.
}
