% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_gsea}
\alias{run_gsea}
\title{Run GSEA using HALLMARK, GO or KEGG gene sets}
\usage{
run_gsea(
  res_df,
  database = c("HALLMARK", "GO", "KEGG"),
  stat = c("log2FoldChange", "stat"),
  seed = NULL
)
}
\arguments{
\item{res_df}{Data frame. Data frame of results generated from the DESeq2 pipeline. Rownames must be ENSEMBL IDs.}

\item{database}{Character. "HALLMARK", "GO" or "KEGG" defining what terms to use for GSEA.}

\item{stat}{Character. "log2FoldChange" or "stat" to define which values are used for GSEA.}

\item{seed}{Number. Any number which will be used as the seed to ensure reproducibility. Default is NULL.}
}
\value{
GSEA results.
}
\description{
This function runs GSEA using clusterProfiler using HALLMARK,
GO or KEGG gene sets/terms based on the database provided.
It is possible to use the log2FoldChange or Wald statistic to perform this.
There is also an option to set a seed to ensure reproducibility of results.
}
