% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_metadata.R
\name{subset_of_interest}
\alias{subset_of_interest}
\title{Subsetting samples of interest for condition-wise analysis}
\usage{
subset_of_interest(sample_list, count_data, metadata, matching_sample_names)
}
\arguments{
\item{sample_list}{Character vector. A list of specific samples by which analysis will be undertaken (e.g. all control & treated in one cell line).}

\item{count_data}{Data frame. The cleaned count data frame from the match_annotations function.}

\item{metadata}{Data frame. The aligned metadata from the match_annotations function.}

\item{matching_sample_names}{Character. The column within the metadata data frame by which samples will be subsetted (i.e. cleaned sample names/clean_col).}
}
\value{
A named list containing the subsetted & matched counts data frame & metadata.
}
\description{
This function allows for the user to subset their counts data by condition/sample name-
-by providing the function a character vector of the samples names of interest.
It is able to match the data with the metadata & order it so that all data are in order.
}
