% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heatmap.R
\name{generate_heatmap_of_interest}
\alias{generate_heatmap_of_interest}
\title{Heatmap generation based on pathway of interest}
\usage{
generate_heatmap_of_interest(
  gsea_res,
  deseq2_object,
  transformation = c("vst", "rlog"),
  annotations,
  pathway_of_interest,
  filename,
  width = 10,
  height = 10
)
}
\arguments{
\item{gsea_res}{Data frame. Results from GSEA.}

\item{deseq2_object}{A DESeq2Results object.}

\item{transformation}{Character. "vst" or "rlog" depending on user intent & sample size. Blind is set to FALSE.}

\item{annotations}{Character or character vector. A single or multiple variables by which the user wishes to annotate samples by (e.g. c("Treatment", "Harvest_Time"))}

\item{pathway_of_interest}{Character. The pathway of interest outputted from GSEA analysis (e.g. "KRAS_SIGNALLING_UP").}

\item{filename}{Character. A string depicting the file name of the heatmap image that the user wishes to save (e.g. sample1_heatmap.png)}

\item{width}{Numerical. Width of resulting heatmap image. Default is 10.}

\item{height}{Numerical. Height of resulting heatmap image. Default is 10.}
}
\value{
Heatmap.
}
\description{
This function generates a heatmap, using pheatmap, with absolute vst or rlog values from the DESeq2Results object for
particular pathways of interest as outputted from GSEA results. All gene names must be ENSEMBL IDs.
}
